removed valid_move function
Most time valid_move was used the calling piece of code needed to know
the result of one of the inner if statements. But it can't so the
calling code would just use another if statement. Effectively the code
ran twice. Every implementation of valid_move has different
requirements, so it was removed.