added valid_move function
Checks if moving a piece to a new position is valid, accounting for king
checks and implementing special moves like en passant and castling. A
new board is returned where the piece is moved to the new coordinates.
En passant currently does not capture a piece. The en passant issue, and
some organisational changes are next.