Fixed en passant not capturing enemy pawn
When doing an en passant valid_move now captures the enemy pawn. Capture
coordinates were also added to the BoardInfo struct. Capture coordinates
specifiy coordinates that were captured that the moving piece did not
replace (en passant captures). This will allow points to be calculated
for the capture more easily in the future.